Separate modifer classes out into separate files
