+++ Assets/Scripts/Incidents/Contexts/ContextModifiers/BooleanContextModifierCalculator.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace Game.Incidents
+{
+	public class BooleanContextModifierCalculator : ContextModifierCalculator<bool>
+    {
+        public BooleanContextModifierCalculator(string propertyName, Type contextType) : base(propertyName, contextType) { }
+        public override void Setup()
+        {
+            Operators = ExpressionHelpers.BoolOperators;
+        }
+    }
+}
