+++ Assets/Scripts/Incidents/Contexts/ContextModifiers/ContextModifier.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3a483a40be0b1994f9198b6dfb3b2953
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
