+++ Assets/Scripts/Incidents/Contexts/ContextModifiers/IContextModifierCalculator.cs
@@ -1,0 +1,7 @@
+﻿namespace Game.Incidents
+{
+	public interface IContextModifierCalculator
+    {
+        void Calculate(IIncidentContext context);
+    }
+}
