+++ Assets/Scripts/Incidents/Contexts/ContextModifiers/IntegerContextModifierCalculator.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace Game.Incidents
+{
+	public class IntegerContextModifierCalculator : ContextModifierCalculator<int>
+    {
+        public IntegerContextModifierCalculator(string propertyName, Type contextType) : base(propertyName, contextType) { }
+        public override void Setup()
+        {
+            Operators = ExpressionHelpers.IntegerOperators;
+        }
+    }
+}
